Fix test failures caused by django test runner
Recently horizon test configuration was changed to use
the Django default test runner because nose is in maintenance mode
for the past several years [1]. Django testcase and oslotest are
incompatible, so congress_dashboard testcase now fails.

test_fake.py actually tests nothing, so there is no reason
to keep it. It sounds reasonable to drop this file.

[1] horizon commit 1f80d94459856a8c477310cc0fe4b0e165d8c0c1

Change-Id: I0b99e709fda9c933ec229bb1dde161cf5bea319b